junção das funções do firebasekitchen no firebasefunctions e alteração nos imports
